{"schema":"decodex/commit/1","summary":"Extract capture host input routing","authority":"manual"}
